requirements.txt
aiogram==3.15.0
aiohttp>=3.9.0,<3.11
asyncpg==0.30.0
sqlalchemy==2.0.36
alembic==1.14.0
qrcode==8.2
Pillow==11.3.0
python-dotenv==1.0.1
redis==5.2.1
